requirements.txt
nltk
pandas
numpy
transformers
openai
python-dotenv
python-docx
pandas
reportlab
fpdf2